requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
flask_sqlalchemy==3.1.1
sqlalchemy_serializer==1.4.22
PyYAML==6.0.2
